+++ bill/charges.go
@@ -194,8 +194,12 @@ func calculateChargeSum(charges []*Charge, cur currency.Code) *num.Amount {
 }
 
 func (m *Charge) round(cur currency.Code) {
-	cd := cur.Def()
-	m.Amount = cd.Rescale(m.Amount)
+	// Default round to currency, or use base if present
+	e := cur.Def().Subunits
+	if m.Base != nil {
+		e = m.Base.Exp()
+	}
+	m.Amount = m.Amount.RescaleDown(e)
 }
 
 func roundCharges(lines []*Charge, cur currency.Code) {